chore: Move extensions/pkg/util/cloudprofile to pkg/utils/gardener/cloudprofile
