+++ pkg/utils/gardener/cloudprofile_test.go
@@ -697,4 +697,98 @@ var _ = Describe("CloudProfile", func() {
 			})
 		})
 	})
+
+	Describe("ImagesContext", func() {
+		Describe("#NewCoreImagesContext", func() {
+			It("should successfully construct an ImagesContext from core.MachineImage slice", func() {
+				imagesContext := gardenerutils.NewCoreImagesContext([]core.MachineImage{
+					{Name: "image-1", Versions: []core.MachineImageVersion{
+						{ExpirableVersion: core.ExpirableVersion{Version: "1.0.0"}},
+						{ExpirableVersion: core.ExpirableVersion{Version: "2.0.0"}},
+					}},
+					{Name: "image-2", Versions: []core.MachineImageVersion{
+						{ExpirableVersion: core.ExpirableVersion{Version: "3.0.0"}},
+					}},
+				})
+
+				i, exists := imagesContext.GetImage("image-1")
+				Expect(exists).To(BeTrue())
+				Expect(i.Name).To(Equal("image-1"))
+				Expect(i.Versions).To(ConsistOf(
+					core.MachineImageVersion{ExpirableVersion: core.ExpirableVersion{Version: "1.0.0"}},
+					core.MachineImageVersion{ExpirableVersion: core.ExpirableVersion{Version: "2.0.0"}},
+				))
+
+				i, exists = imagesContext.GetImage("image-2")
+				Expect(exists).To(BeTrue())
+				Expect(i.Name).To(Equal("image-2"))
+				Expect(i.Versions).To(ConsistOf(
+					core.MachineImageVersion{ExpirableVersion: core.ExpirableVersion{Version: "3.0.0"}},
+				))
+
+				i, exists = imagesContext.GetImage("image-99")
+				Expect(exists).To(BeFalse())
+				Expect(i.Name).To(Equal(""))
+				Expect(i.Versions).To(BeEmpty())
+
+				v, exists := imagesContext.GetImageVersion("image-1", "1.0.0")
+				Expect(exists).To(BeTrue())
+				Expect(v).To(Equal(core.MachineImageVersion{ExpirableVersion: core.ExpirableVersion{Version: "1.0.0"}}))
+
+				v, exists = imagesContext.GetImageVersion("image-1", "99.0.0")
+				Expect(exists).To(BeFalse())
+				Expect(v).To(Equal(core.MachineImageVersion{}))
+
+				v, exists = imagesContext.GetImageVersion("image-99", "99.0.0")
+				Expect(exists).To(BeFalse())
+				Expect(v).To(Equal(core.MachineImageVersion{}))
+			})
+		})
+
+		Describe("#NewV1beta1ImagesContext", func() {
+			It("should successfully construct an ImagesContext from v1beta1.MachineImage slice", func() {
+				imagesContext := gardenerutils.NewV1beta1ImagesContext([]gardencorev1beta1.MachineImage{
+					{Name: "image-1", Versions: []gardencorev1beta1.MachineImageVersion{
+						{ExpirableVersion: gardencorev1beta1.ExpirableVersion{Version: "1.0.0"}},
+						{ExpirableVersion: gardencorev1beta1.ExpirableVersion{Version: "2.0.0"}},
+					}},
+					{Name: "image-2", Versions: []gardencorev1beta1.MachineImageVersion{
+						{ExpirableVersion: gardencorev1beta1.ExpirableVersion{Version: "3.0.0"}},
+					}},
+				})
+
+				i, exists := imagesContext.GetImage("image-1")
+				Expect(exists).To(BeTrue())
+				Expect(i.Name).To(Equal("image-1"))
+				Expect(i.Versions).To(ConsistOf(
+					gardencorev1beta1.MachineImageVersion{ExpirableVersion: gardencorev1beta1.ExpirableVersion{Version: "1.0.0"}},
+					gardencorev1beta1.MachineImageVersion{ExpirableVersion: gardencorev1beta1.ExpirableVersion{Version: "2.0.0"}},
+				))
+
+				i, exists = imagesContext.GetImage("image-2")
+				Expect(exists).To(BeTrue())
+				Expect(i.Name).To(Equal("image-2"))
+				Expect(i.Versions).To(ConsistOf(
+					gardencorev1beta1.MachineImageVersion{ExpirableVersion: gardencorev1beta1.ExpirableVersion{Version: "3.0.0"}},
+				))
+
+				i, exists = imagesContext.GetImage("image-99")
+				Expect(exists).To(BeFalse())
+				Expect(i.Name).To(Equal(""))
+				Expect(i.Versions).To(BeEmpty())
+
+				v, exists := imagesContext.GetImageVersion("image-1", "1.0.0")
+				Expect(exists).To(BeTrue())
+				Expect(v).To(Equal(gardencorev1beta1.MachineImageVersion{ExpirableVersion: gardencorev1beta1.ExpirableVersion{Version: "1.0.0"}}))
+
+				v, exists = imagesContext.GetImageVersion("image-1", "99.0.0")
+				Expect(exists).To(BeFalse())
+				Expect(v).To(Equal(gardencorev1beta1.MachineImageVersion{}))
+
+				v, exists = imagesContext.GetImageVersion("image-99", "99.0.0")
+				Expect(exists).To(BeFalse())
+				Expect(v).To(Equal(gardencorev1beta1.MachineImageVersion{}))
+			})
+		})
+	})
 })
